+++ core/src/io/github/nemo/cap/UpdateMetadata.java
@@ -15,9 +15,9 @@ public final class UpdateMetadata {
     public static JSONObject evaluate(JSONObject data,int current) throws Exception {
         Object code=data.get("versionCode");String version=data.getString("version");
         Failure.require(code instanceof Integer && (Integer)code>0 && version.matches("v[0-9]+\\.[0-9]+\\.[0-9]+"),"INVALID_UPDATE_METADATA","Invalid version metadata");
-        String release="https://github.com/nemo201104/Clone-App-Profile/releases/";
+        String release="https://github.com/nemoforge/Clone-App-Profile/releases/";
         Failure.require(data.getString("zipUrl").equals(release+"download/"+version+"/Clone-App-Profile-"+version+".zip")
-            && data.getString("changelog").equals("https://raw.githubusercontent.com/nemo201104/Clone-App-Profile/"+version+"/CHANGELOG.md"),"INVALID_UPDATE_METADATA","Unexpected release URL");
+            && data.getString("changelog").equals("https://raw.githubusercontent.com/nemoforge/Clone-App-Profile/"+version+"/CHANGELOG.md"),"INVALID_UPDATE_METADATA","Unexpected release URL");
         return Json.obj("state",data.getInt("versionCode")>current?"UPDATE_AVAILABLE":"UP_TO_DATE","installedVersionCode",current,"metadata",data);
     }
 }